Refactor of Instantiation (#20)
* implementation of new instantiation approach (mainly for steps)

+++ src/TeaPie/Pipelines/TemporaryFolder/CleanUpTemporaryFolderStep.cs
@@ -5,12 +5,13 @@ namespace TeaPie.Pipelines.TemporaryFolder;
 
 internal sealed class CleanUpTemporaryFolderStep : IPipelineStep
 {
-    private CleanUpTemporaryFolderStep() { }
-
-    public static CleanUpTemporaryFolderStep Create() => new();
-
     public async Task Execute(ApplicationContext context, CancellationToken cancellationToken = default)
     {
+        if (!Directory.Exists(context.TempFolderPath))
+        {
+            throw new DirectoryNotFoundException($"Directory on path {context.TempFolderPath} does not exist.");
+        }
+
         Directory.Delete(context.TempFolderPath, true);
         await Task.CompletedTask;
